+++ src/components/SideBar/Notifications.tsx
@@ -9,10 +9,7 @@ import { UserInfo } from "../../store/types/userInfoType";
 import { NotificationActions } from "../../store/actions/notificationActions";
 import { firebase } from "../../utils/firebase";
 import { NoSidebarDataText } from "./FollowList";
-import {
-  faArrowUpRightFromSquare,
-  faCircleXmark,
-} from "@fortawesome/free-solid-svg-icons";
+import { faCircleXmark } from "@fortawesome/free-solid-svg-icons";
 const NoticeWrapper = styled.div`
   width: 100%;
   height: 300px;
@@ -22,7 +19,7 @@ const NoticeWrapper = styled.div`
   background-color: #fff;
 `;
 interface NoticeProps {
-  show: boolean;
+  $read: boolean;
 }
 const StyleWrapper = styled.div`
   width: 230px;
@@ -37,18 +34,14 @@ const Notice = styled.div<NoticeProps>`
   display: flex;
   align-items: center;
   justify-content: space-between;
+  background-color: ${(props) =>
+    props.$read ? "#fff" : "rgba(92, 131, 111, 0.2)"};
   margin-left: 8px;
   transition: 0.25s;
   &:hover {
     text-decoration: underline;
     transition: 0.25s;
   }
-`;
-const UnreadMark = styled.div`
-  width: 10px;
-  height: 10px;
-  border-radius: 50%;
-  background-color: #5c836f;
 `;
 const NoticeText = styled.div`
   font-size: 14px;
@@ -100,7 +93,7 @@ const Notifications = ({ notices, followInfos }: NotificationsProps) => {
       type: NotificationActions.UPDATE_READ_STATUS,
       payload: { noticeId },
     });
-    await firebase.updateReadStatus(userId, noticeId);
+    await firebase.updateNoticeReadStatus(userId, noticeId);
   }
   return (
     <NoticeWrapper>
@@ -109,7 +102,7 @@ const Notifications = ({ notices, followInfos }: NotificationsProps) => {
           return (
             <Notice
               key={note.noticeId}
-              show={note.read}
+              $read={note.read}
               onClick={() => {
                 if (!note.read) {
                   changeReadStatus(note.noticeId);
@@ -117,7 +110,6 @@ const Notifications = ({ notices, followInfos }: NotificationsProps) => {
               }}
             >
               <StyleWrapper>
-                {!note.read && <UnreadMark />}
                 {note.type === "1" && (
                   <NoticeText
                     onClick={() => navigate(`/profile/${note.userId}`)}
@@ -128,7 +120,6 @@ const Notifications = ({ notices, followInfos }: NotificationsProps) => {
                     onClick={() => navigate(`/forum/${note.postId}`)}
                   >{`${findUserName(note.userId)}${noticeMsg2}`}</NoticeText>
                 )}
-                <StyledFontAwesomeIcon icon={faArrowUpRightFromSquare} />
               </StyleWrapper>
               <CloseFontAwesomeIcon
                 icon={faCircleXmark}
