+++ src/utils/firebase.ts
@@ -26,7 +26,7 @@ import { getStorage } from "firebase/storage";
 import { getAuth } from "firebase/auth";
 import { UserInfo } from "../store/types/userInfoType";
 import { message } from "../components/Chatroom/Chatroom";
-import { Comment, Post } from "../pages/Forum/ForumPost";
+import { Comment, Post } from "../pages/Forum/ForumPost/ForumPost";
 import { PlantCard } from "../store/types/plantCardType";
 import { Note } from "../store/types/notificationType";
 import { unixTimeToString } from "./helpers";
@@ -65,6 +65,12 @@ const firebase = {
     const docSnapshot = await getDoc(docRef);
     return docSnapshot;
   },
+  async getUsers(idList: string[]) {
+    if (!idList.length) return;
+    const q = query(users, where("userId", "in", idList));
+    const querySnapshot = await getDocs(q);
+    return querySnapshot;
+  },
   async updateUserPhoto(id: string, url: string) {
     const docRef = doc(users, id);
     await updateDoc(docRef, { photoUrl: url });
@@ -72,133 +78,6 @@ const firebase = {
   async updateUserName(id: string, name: string) {
     const docRef = doc(users, id);
     await updateDoc(docRef, { userName: name });
-  },
-  async checkChatroom(users: string[]) {
-    const usersCopy = [...users];
-    const q = query(
-      chatrooms,
-      where("users", "in", [users, usersCopy.reverse()])
-    );
-    const querySnapshot = await getDocs(q);
-    if (querySnapshot.empty) {
-      await addDoc(chatrooms, {
-        users: users,
-        msgs: [],
-      });
-    }
-  },
-  async storeChatroomData(users: string[], msg: message) {
-    const usersCopy = [...users];
-    const q = query(
-      chatrooms,
-      where("users", "in", [users, usersCopy.reverse()])
-    );
-    const querySnapshot = await getDocs(q);
-    if (querySnapshot.empty) {
-      const room = doc(chatrooms);
-      await setDoc(room, { users: users, msgs: [msg] });
-    } else {
-      querySnapshot.forEach(async (docData) => {
-        const docRef = doc(chatrooms, docData.id);
-        const docMsgs = docData.data().msgs;
-        await updateDoc(docRef, { msgs: [...docMsgs, msg] });
-      });
-    }
-  },
-  async changePlantOwner(cardId: string, newOwnerId: string) {
-    const docRef = doc(cards, cardId);
-    await updateDoc(docRef, { ownerId: newOwnerId });
-  },
-  async addPost(postData: {
-    title: string;
-    content: string;
-    authorId: string;
-    type: string;
-  }) {
-    const post = doc(posts);
-    await setDoc(post, {
-      ...postData,
-      postId: post.id,
-      createdTime: serverTimestamp(),
-    });
-    return post.id;
-  },
-  async getPostData(postId: string) {
-    const docRef = doc(posts, postId);
-    const docSnapshot = await getDoc(docRef);
-    return docSnapshot;
-  },
-  async saveEditPost(postId: string, post: Post) {
-    const docRef = doc(posts, postId);
-    const docSnapshot = await setDoc(docRef, {
-      ...post,
-      createdTime: serverTimestamp(),
-    });
-    return docSnapshot;
-  },
-  async deletePost(postId: string) {
-    const docRef = doc(posts, postId);
-    const docSnapshot = await deleteDoc(docRef);
-    return docSnapshot;
-  },
-  async getPosts() {
-    const q = query(posts, orderBy("createdTime", "desc"));
-    const querySnapshot = getDocs(q);
-    return querySnapshot;
-  },
-  async getLatestPosts() {
-    const q = query(posts, orderBy("createdTime", "asc"), limit(5));
-    const querySnapshot = getDocs(q);
-    return querySnapshot;
-  },
-  async saveComment(
-    postId: string,
-    comment: { content: string; authorId: string; createdTime: number }
-  ) {
-    const docRef = doc(posts, postId);
-    await updateDoc(docRef, { comments: arrayUnion(comment) });
-  },
-  async saveEditComment(postId: string, comments: Comment[]) {
-    const docRef = doc(posts, postId);
-    await updateDoc(docRef, { comments: comments });
-  },
-  async getCardData(cardId: string) {
-    const docRef = doc(cards, cardId);
-    const docSnapshot = await getDoc(docRef);
-    return docSnapshot;
-  },
-  async getCards(cardIds: string[]) {
-    if (cardIds.length === 0) return;
-    const q = query(cards, where("cardId", "in", cardIds));
-    const querySnapshot = await getDocs(q);
-    return querySnapshot;
-  },
-  async getUserCards(ownerId: string) {
-    const q = query(
-      cards,
-      where("ownerId", "==", ownerId),
-      orderBy("createdTime", "asc")
-    );
-    const querySnapshot = await getDocs(q);
-    return querySnapshot;
-  },
-  async getFavCards() {
-    const q = query(cards, orderBy("followers", "desc"), limit(10));
-    const querySnapshot = await getDocs(q);
-    return querySnapshot;
-  },
-  async addCard(data: PlantCard) {
-    const newCard = doc(cards);
-    await setDoc(newCard, {
-      ...data,
-      cardId: newCard.id,
-      createdTime: serverTimestamp(),
-    });
-    return newCard.id;
-  },
-  async editCard(cardId: string, data: PlantCard) {
-    const docRef = doc(cards, cardId);
-    await setDoc(docRef, data);
   },
   async addFollowList(selfId: string, followId: string) {
     const docRef = doc(users, selfId);
@@ -235,17 +114,6 @@ const firebase = {
       followers: arrayRemove(selfId),
     });
     Promise.all([updateFollowList, removeFollower]);
-  },
-  async getUsers(idList: string[]) {
-    if (!idList.length) return;
-    const q = query(users, where("userId", "in", idList));
-    const querySnapshot = await getDocs(q);
-    return querySnapshot;
-  },
-  async getChatrooms(userId: string) {
-    const q = query(chatrooms, where("users", "array-contains", userId));
-    const querySnapshot = await getDocs(q);
-    return querySnapshot;
   },
   async emitNotice(
     userId: string,
@@ -284,42 +152,9 @@ const firebase = {
     const docRef = doc(db, "users", userId, "notices", noticeId);
     await deleteDoc(docRef);
   },
-  async updateReadStatus(userId: string, noticeId: string) {
+  async updateNoticeReadStatus(userId: string, noticeId: string) {
     const docRef = doc(db, "users", userId, "notices", noticeId);
     await updateDoc(docRef, { read: true });
-  },
-  async addFavCard(userId: string, cardId: string) {
-    const docRef = doc(users, userId);
-    const cardDocRef = doc(cards, cardId);
-    const updateUser = updateDoc(docRef, { favoriteCards: arrayUnion(cardId) });
-    const updateCard = updateDoc(cardDocRef, { followers: increment(1) });
-    await Promise.all([updateUser, updateCard]);
-  },
-  async removeFavCard(userId: string, cardId: string) {
-    const docRef = doc(users, userId);
-    const cardDocRef = doc(cards, cardId);
-    const updateUser = updateDoc(docRef, {
-      favoriteCards: arrayRemove(cardId),
-    });
-    const updateCard = updateDoc(cardDocRef, { followers: increment(-1) });
-    await Promise.all([updateUser, updateCard]);
-  },
-  async getDiary(diaryId: string) {
-    const docRef = doc(diaries, diaryId);
-    const docSnapshot = await getDoc(docRef);
-    return docSnapshot;
-  },
-  async saveDiary(diaryId: string, page: string) {
-    const docRef = doc(diaries, diaryId);
-    const docSnapshot = await getDoc(docRef);
-    if (!docSnapshot.exists()) {
-      await setDoc(docRef, { pages: [] });
-    }
-    await updateDoc(docRef, { pages: arrayUnion(page) });
-  },
-  async saveEditDiary(diaryId: string, currentDiaries: string[]) {
-    const docRef = doc(diaries, diaryId);
-    await setDoc(docRef, { pages: currentDiaries });
   },
   async getEvent(docName: string, id: string) {
     const activitiesRef = collection(db, "users", id, "activities");
@@ -345,9 +180,6 @@ const firebase = {
       await updateDoc(docRef, { fertilizing: arrayUnion(...cardIds) });
     }
   },
-  async deleteCard(cardId: string) {
-    await deleteDoc(doc(db, "cards", cardId));
-  },
   async getGallery(userId: string) {
     const docRef = doc(users, userId);
     const docSnapshot = await getDoc(docRef);
@@ -361,24 +193,167 @@ const firebase = {
     const docRef = doc(users, userId);
     await updateDoc(docRef, { gallery: arrayRemove(link) });
   },
-  async searchSpecies(input: string) {
-    const q = query(species, where("NAME", "==", input));
+  async addFavCard(userId: string, cardId: string) {
+    const docRef = doc(users, userId);
+    const cardDocRef = doc(cards, cardId);
+    const updateUser = updateDoc(docRef, { favoriteCards: arrayUnion(cardId) });
+    const updateCard = updateDoc(cardDocRef, { followers: increment(1) });
+    await Promise.all([updateUser, updateCard]);
+  },
+  async removeFavCard(userId: string, cardId: string) {
+    const docRef = doc(users, userId);
+    const cardDocRef = doc(cards, cardId);
+    const updateUser = updateDoc(docRef, {
+      favoriteCards: arrayRemove(cardId),
+    });
+    const updateCard = updateDoc(cardDocRef, { followers: increment(-1) });
+    await Promise.all([updateUser, updateCard]);
+  },
+  //chatroom
+  async checkChatroom(users: string[]) {
+    const usersCopy = [...users];
+    const q = query(
+      chatrooms,
+      where("users", "in", [users, usersCopy.reverse()])
+    );
+    const querySnapshot = await getDocs(q);
+    if (querySnapshot.empty) {
+      await addDoc(chatrooms, {
+        users: users,
+        msgs: [],
+      });
+    }
+  },
+  async storeChatroomData(users: string[], msg: message) {
+    const usersCopy = [...users];
+    const q = query(
+      chatrooms,
+      where("users", "in", [users, usersCopy.reverse()])
+    );
+    const querySnapshot = await getDocs(q);
+    if (querySnapshot.empty) {
+      const room = doc(chatrooms);
+      await setDoc(room, { users: users, msgs: [msg] });
+    } else {
+      querySnapshot.forEach(async (docData) => {
+        const docRef = doc(chatrooms, docData.id);
+        const docMsgs = docData.data().msgs;
+        await updateDoc(docRef, { msgs: [...docMsgs, msg] });
+      });
+    }
+  },
+  async getChatrooms(userId: string) {
+    const q = query(chatrooms, where("users", "array-contains", userId));
     const querySnapshot = await getDocs(q);
     return querySnapshot;
   },
-  async uploadFile(file: File) {
-    const galleryId = uuidv4();
-    const storageRef = ref(storage, `${galleryId}`);
-    await uploadBytes(storageRef, file);
-    const dowloadLink = await getDownloadURL(storageRef);
-    return dowloadLink;
-  },
-  async checkOwner(cardId: string) {
+  //cards coolection
+  async changePlantOwner(cardId: string, newOwnerId: string) {
     const docRef = doc(cards, cardId);
+    await updateDoc(docRef, { ownerId: newOwnerId });
+  },
+  async getCardsByIds(cardIds: string[]) {
+    if (cardIds.length === 0) return;
+    const q = query(cards, where("cardId", "in", cardIds));
+    const querySnapshot = await getDocs(q);
+    return querySnapshot;
+  },
+  async getUserCards(ownerId: string) {
+    const q = query(
+      cards,
+      where("ownerId", "==", ownerId),
+      orderBy("createdTime", "asc")
+    );
+    const querySnapshot = await getDocs(q);
+    return querySnapshot;
+  },
+  async getPopularCards() {
+    const q = query(cards, orderBy("followers", "desc"), limit(10));
+    const querySnapshot = await getDocs(q);
+    return querySnapshot;
+  },
+  async addCard(data: PlantCard) {
+    const newCard = doc(cards);
+    await setDoc(newCard, {
+      ...data,
+      cardId: newCard.id,
+      createdTime: serverTimestamp(),
+    });
+    return newCard.id;
+  },
+  async editCard(cardId: string, data: PlantCard) {
+    const docRef = doc(cards, cardId);
+    await setDoc(docRef, data);
+  },
+  async deleteCard(cardId: string) {
+    await deleteDoc(doc(db, "cards", cardId));
+  },
+  //post collection
+  async addPost(postData: {
+    title: string;
+    content: string;
+    authorId: string;
+    type: string;
+  }) {
+    const post = doc(posts);
+    await setDoc(post, {
+      ...postData,
+      postId: post.id,
+      createdTime: serverTimestamp(),
+    });
+    return post.id;
+  },
+  async getPostData(postId: string) {
+    const docRef = doc(posts, postId);
     const docSnapshot = await getDoc(docRef);
-    if (docSnapshot.exists()) {
-      return docSnapshot.data().ownerId;
+    return docSnapshot;
+  },
+  async saveEditPost(postId: string, post: Post) {
+    const docRef = doc(posts, postId);
+    const docSnapshot = await setDoc(docRef, {
+      ...post,
+      createdTime: serverTimestamp(),
+    });
+    return docSnapshot;
+  },
+  async deletePost(postId: string) {
+    const docRef = doc(posts, postId);
+    const docSnapshot = await deleteDoc(docRef);
+    return docSnapshot;
+  },
+  async getPosts() {
+    const q = query(posts, orderBy("createdTime", "desc"));
+    const querySnapshot = getDocs(q);
+    return querySnapshot;
+  },
+  async saveComment(
+    postId: string,
+    comment: { content: string; authorId: string; createdTime: number }
+  ) {
+    const docRef = doc(posts, postId);
+    await updateDoc(docRef, { comments: arrayUnion(comment) });
+  },
+  async saveEditComment(postId: string, comments: Comment[]) {
+    const docRef = doc(posts, postId);
+    await updateDoc(docRef, { comments: comments });
+  },
+  // diary collection
+  async getDiary(diaryId: string) {
+    const docRef = doc(diaries, diaryId);
+    const docSnapshot = await getDoc(docRef);
+    return docSnapshot;
+  },
+  async saveDiary(diaryId: string, page: string) {
+    const docRef = doc(diaries, diaryId);
+    const docSnapshot = await getDoc(docRef);
+    if (!docSnapshot.exists()) {
+      await setDoc(docRef, { pages: [] });
     }
+    await updateDoc(docRef, { pages: arrayUnion(page) });
+  },
+  async saveEditDiary(diaryId: string, currentDiaries: string[]) {
+    const docRef = doc(diaries, diaryId);
+    await setDoc(docRef, { pages: currentDiaries });
   },
   async checkDiaryExistence(id: string) {
     const docRef = doc(diaries, id);
@@ -393,18 +368,18 @@ const firebase = {
     const result = await Promise.all(promises);
     return result;
   },
+  async searchSpecies(input: string) {
+    const q = query(species, where("NAME", "==", input));
+    const querySnapshot = await getDocs(q);
+    return querySnapshot;
+  },
+  async uploadFile(file: File) {
+    const galleryId = uuidv4();
+    const storageRef = ref(storage, `${galleryId}`);
+    await uploadBytes(storageRef, file);
+    const dowloadLink = await getDownloadURL(storageRef);
+    return dowloadLink;
+  },
 };
 
-export {
-  app,
-  db,
-  storage,
-  auth,
-  users,
-  cards,
-  posts,
-  species,
-  chatrooms,
-  diaries,
-  firebase,
-};
+export { db, chatrooms, auth, firebase };
